refactor: keep session recovery text independent of delivery runtime
