+++ src/auto-reply/reply/pending-final-delivery-state.ts
@@ -1,0 +1,42 @@
+// Recovery admission must not load outbound normalization or provider runtime to inspect text.
+import type { SessionEntry } from "../../config/sessions/types.js";
+import {
+  isSilentReplyPayloadText,
+  isSilentReplyText,
+  SILENT_REPLY_TOKEN,
+  startsWithSilentToken,
+  stripLeadingSilentToken,
+  stripSilentToken,
+} from "../tokens.js";
+import { stripInternalMetadataForDisplay } from "./display-text-sanitize.js";
+
+// A delivered or discarded final must lose the whole record. Keeping this list
+// centralized prevents new ownership fields from leaving a phantom pending delivery.
+export const PENDING_FINAL_DELIVERY_CLEAR_PATCH = {
+  pendingFinalDelivery: undefined,
+} as const satisfies Partial<SessionEntry>;
+
+/** Sanitizes pending final delivery text before channel-visible output. */
+export function sanitizePendingFinalDeliveryText(text: string): string {
+  let stripped = stripInternalMetadataForDisplay(text).trim();
+  if (isSilentReplyPayloadText(stripped, SILENT_REPLY_TOKEN)) {
+    return "";
+  }
+  if (stripped && !isSilentReplyText(stripped, SILENT_REPLY_TOKEN)) {
+    const hasLeadingSilentToken = startsWithSilentToken(stripped, SILENT_REPLY_TOKEN);
+    if (hasLeadingSilentToken) {
+      stripped = stripLeadingSilentToken(stripped, SILENT_REPLY_TOKEN);
+    }
+    // Remove stray silent tokens only after confirming the payload is not entirely silent.
+    if (
+      hasLeadingSilentToken ||
+      stripped.toLowerCase().includes(SILENT_REPLY_TOKEN.toLowerCase())
+    ) {
+      stripped = stripSilentToken(stripped, SILENT_REPLY_TOKEN);
+    }
+  }
+  if (!stripped.trim()) {
+    return "";
+  }
+  return isSilentReplyPayloadText(stripped, SILENT_REPLY_TOKEN) ? "" : stripped.trim();
+}
